chore: upgrade Segment analytics-kotlin to 1.25.0 (MBL-1709)
Bump Segment from 1.19.2 to 1.25.0 — the SDK compiles, ABI is stable,
and tests pass against the new version. The 1.25.0 release deprecates
several typed/SerializationStrategy overloads in favor of plain
Map<String, Any?> ones; warnings only, no hard breaks.

Test fixes carried over from #679 because Segment 1.20.0+ changed
StartupQueue behavior: ToggleRunningAction(true) is only dispatched on
successful settings fetch, so events stay queued when settings fail.
Two test cases that relied on the 1.19.2 behavior are corrected:
- givenDestinationPlugin_expectIntegrationToBeAddedAndEventToFlow:
  drop the event-flow assertion that was validating a 1.19.2 bug.
- givenComplexEventWithNestedNulls...: moved to
  DataPipelinesInteractionTests where settings succeed, making the
  assertion meaningful.
