+++ buildSrc/src/main/kotlin/io.customer/android/Versions.kt
@@ -30,7 +30,7 @@ object Versions {
     internal const val KOTLINX_SERIALIZATION_JSON = "1.8.1"
     internal const val KOTLIN = "2.1.21"
     internal const val MOCKK = "1.13.13"
-    internal const val SEGMENT = "1.19.2"
+    internal const val SEGMENT = "1.25.0"
     internal const val REDUX_KOTLIN = "0.6.0"
     internal const val ROBOLECTRIC = "4.16"
     internal const val OKHTTP = "4.12.0"
